requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
gunicorn==20.1.0
pydantic==1.10
httpx==0.24.1
pytest==7.4.0
pytest-asyncio==0.21.1

SQLAlchemy==1.4.46
asyncpg==0.27.0
aiosqlite==0.17.0

passlib[argon2]==1.7.4
PyJWT==2.8.0
python-dotenv==1.0.0
redis[hiredis]==5.0.1